Parallelize Instagram synclets
This lets Instagram synclets run totally independently of one another. I've
updated the config names to be simpler, since the variable names don't need
their own scoping anymore.

This will cause the media and feed synclets to run a full rescan since they're
effectively losing their configs, but that should be fine. Instagram is the last
service with a relatively large number of profiles, so it's probably not worth
running migrations.