Use the time a Flickr photo was taken as its `at`
Not all photos include a `datetaken` (depends on the EXIF), but if it's there,
we use it as the photo's `at` timestamp. If Flickr reports a timezone for the
user, we use it to parse the time, otherwise we assume UTC.

In the event `datetaken` is missing, we fall back to `dateupload`, which is
always seconds since epoch GMT.

Because the timezone is stored in the user's profile data, I move figuring out
`at` from `map` into the `photos` synclet. `map` doesn't have access to the
profile, so it couldn't be done there.

This effects the IJOD entry hash, so we'll need to wipe and recrawl Flickr data.
Flickr usage is light, and we've notified the app using it of the change, so it
shouldn't be a big deal. The most someone will see is a potential reordering of
photos returned from the API.