Support sandboxed synclets in testSynclet
Sandboxed synclet configs aren't compatible with unsandboxed ones because of the
namespacing that happens, so they would either run from a blank slate every
time, or bail in the event that they needed the `all` config section to
communicate. Now we check whether the synclets are sandboxed and wiggle the
config into the right shape.

The returned `pi` that we use to support paging needs to do the merging that
happens in the real world, because synclets don't always hand back their full
config object - only updates. We're not quite doing that yet, but none of the
synclets I've been testing have cared, so I'm putting it off. Setting `pi =
data` is almost certainly totally wrong because it throws out the `auth` object,
but that's how it was and I don't have a use case to fix it yet.

Really what needs to happen is that we refactor `taskman` so that we don't have
to redo so much of this work. This parallel harness is only getting larger and
has already started to diverge.