Add script to verify contents of the profile store
This runs through each profile ID in the `Profiles` table and checks whether
it's in the KV store. Yes, it loads all the profile IDs in one DB call - since
it's the primary key, that comes back real quick, and it's only a few MB. Didn't
feel worth spending the time on paging.