Implement remote profile fetching
Now if the profile has a `pod`, it gets fetched through the `podClient`, which
I've built some utilities into to simplify HTTP auth'd requests and routing for
us. The `podService` has one endpoint now at `/profile_data` that returns the
object containing `auth` and `config`.

Because I just have the one DB running locally, there was fun looping behavior
when the pod tried to remotely request something from itself (because the `pod`
column contained something). So for now, if we're running under the `pod` role,
we will *always* do a local fetch. We can remove this restriction later if we
want pods to start routing between each other, but we'll need to teach them
which pod they are to avoid the loop.

To run locally, I've added `pod1.localhost` to `/etc/hosts`.