Add tests for local profileManager behavior
This tests what happens on the `allGet` paths that don't require talking to a
pod:
- Getting data when the Profile exists and has no `pod` ID
- Creating a Profile when we are a pod

To mock `auth` and `config` data, I made `profileManager` export `_kvstore` so
we can insert the data we want to test for.

Though `fakeweb` isn't used yet, it ensures that no network traffic is attempted
during these tests, for example if `profileManager` were to try to talk to a
pod.