Don't link to an Account when Profile save fails
We were creating an `Accounts` row even if the `Profile` wasn't successfully
saved on the pod (eg, if the pod was down). Now we run `profileManager.allGet`
first-off simply to make sure that happens, and bail if something goes wrong.
I'm sad that we have to run `profileManager.authSet` later and can't just do it
here, but we don't necessarily have an `accountID` yet, and we need that to set
some data in the `auth` blob.