Fix oEmbed pump cache
We've been passing IJOD an IDR object when it expects a string, which makes it
immediately return an error. As a result, the URL cacheing we've meant to have
on oEmbed lookups has not actually been working at all.

I'm making this change here instead of on `master` because we're merging soon
and it would only cause a conflict.