Bottleneck profileManager.authSet through allSet
Besides greatly simplifying the code, this puts all merging logic in one place,
which will allow us to build just one pod endpoint that receives a mergeable
object. Otherwise, we would have been forced to fetch the data from the pod,
merge locally, the send it back to the pod.

Some of the simplification emerged from the fact that even though `authSet`
treated `app` as optional, we always call with one. I rearranged the arguments
to help indicate that it's no longer optional, and to put the effective key
(`[pid, app]`) together.