Add endpoint to clear pod profile configs
Rounding out our CRUD on remote profiles. This will send an HTTP `DELETE` to the
pod containing a profile we want to clear config data on. Though philosophically
that should delete the whole profile, we never actually want that to happen. We
would lose track of a lot of data, and wiping `auth` would break syncing for all
apps authorized to see that profile. So `DELETE /profile` only clears the
`config` object, just like `reset` has always done locally.