Make acl fetch apps from the nexus
`taskman` determine which synclets to run based on the app data, which isn't
stored in pods. `acl.getApp` appears to be the single thing it needs to do its
work, so we're starting with that.

`nexusClient` exposes a function to get a list of apps, but it works the same
way as places that are fetching many already, so I'm not worrying about those
for now. We're likely going to want to fetch a bundle of apps in one request
soon anyway.

To decide whether `acl` should ask the nexus or look locally (because we _are_
the nexus), the simplest thing is to just tell `acl` what role we're running
under and switch on that. A result is that you **must** be running a nexus now
in order for anything to work. Because for the moment `nexus` and `apihost` are
sharing DB configs, writing the app data doesn't have to change yet.