Load peer profiles from the nexus
Since your friends' profiles might not be on this pod, we need to talk to the
nexus to get all the accounts associated with your associates. The query is
remarkably similar to one in `acl`, so I moved it there and added endpoints in
`nexus{Client,Service}`.

Since we're not creating anything, I changed the `POST` in the client method to
a `GET`, but because we might have a long list of IDs, I continued passing them
in the request body. Both Node processes are happy accepting this even though
it's not totally to HTTP spec (tested with 100,000 IDs). If there's ever a
problem, we can simple switch to `POST`ing.