Make Instagram synclets page properly
These were taking up the worker process for upwards of 150 seconds during
recursion on long feeds. Now they page like everyone else, and I upped the page
size from `20` to `100` to be more efficient about it in general.

There's some fun behavior documented in `feed` regarding instagram's `/feed`
endpoint not accepting a timestamp like the rest of the endpoints. In short, we
get an empty set back if ever we pass a cursor ID that doesn't exist (was
probably deleted). This can only happen if someone deleted the post in between
us receiving the cursor URL from Instagram and the next synclet run, which if
the system is running well will be short. In that event, we'll think we've found
the end of the feed and stop going back in time. So while it's possible for us
to miss chunks in this event, I expect it to be quite rare.