Add localLevel, logger.vital
lconfig.logging.localLevel controls what log messages are prefixed with
LOGLOCAL. It defaults to `debug` so that it doesn't affect your local log
messages. In production it should be set to `vital`, so that `info` and `debug`
messages are prefixed with LOGLOCAL and don't get sent to papertrail.

I also audited all our uses of logger.info and promoted some to logger.vital
since they should be sent to papertrail (even though they aren't warnings).