fix instagram synclet
the instagram synclet wasn't paging properly. This changes it to a
pattern that works better for rescheduling the next page. As a note,
the instagram API doesn't appear to respect the `count` passed in and
instead returns a number > 20 and < `count`.

For the next page, we rely on on the pagination url provided. We
provide the min_timestamp we want to see (if a full sync has
previously been completed) so we stop when instagram stops providing
pagination urls.

We also stop if the min_timestamp is within the current page's max and
min, meaning we should have seen all older items. This is needed to
sync the social feed, which doesn't support min_timestamp.